Adding rough draft of doc on compiling project to executable.
This is incomplete and needs more work before we merge in. I don't know how things work on Linux so I'm leaving that until someone either can help me learn what I need to know to finish that or wants to add them in for me. Those with access to the main repo should be able to push to this fork.

Things I need help with:
* what are file extensions on Linux and Mac OS after compiling
* where will Nvgt generally be on Linux and Mac OS, and how to find the lib folder
* how compiling an nvgt file is done on Linux and Mac OS
* This might be grounds for another issue to discuss better, I can't get sound out of my bgt voiceover script after compiling a test program unless I have phonon.dll. I mentioned such in the doc but it may be unrelated to others. As far as I know, I'm not using a sound environment, just the old bgt-style sound:: object. e.g. sound voSound1; sound voSound2;. There are a few things happening like checking how far voSound is along the duration of play versus how long it is in total. If I try to use it, the game closes after about a second of no sound. It does run until it needs to play the sound though, letting me fill in some edit boxes with the input_box function and such.