Added Link control type to audio form
After nvgt has introduced a function to open a link in the default web browser and users have a choice to use run functioin to run external programs as well, the proper Link type is requires, so users have an alternative to the button control. Currently, this control only introduces a link that gives a semantic link control. If the community agrees, we can add default click behavior to it as well.