Work on UIA support a bit
The first implementation of this was not working correctly because it only attempted to initialize UIA once, potentialy before a window was shown. After that, it assumed it was unavailable. Renaming the NVDAControlerClient dll caused SAPI to be used in the form test instead.

Beyond that a check caused the initialization to abort if com failed to initialize, accepted that something else on the same thread may have initialized com previously and a second valid call after initialization may fail.

So now UIA seems to initialize properly, however nothing speaks. That's where I'll leave off for the moment.