Fix pathfinder memory leak
The pathfinder was generating circular references if a class held an instance to a pathfinder which had a callback function set to a delegate of a method in that class. Implement gc behaviors in the pathfinder to stop this from happening, and fix an invalid invocation of the garbage collector while at it.