Fix bug when resetting active profiler
If reset_profiler was called while profiling was enabled, the callback
would attempt to increase the elapsed time of a non-existent entry
in the cache. This caused uninitialized values to be stored, leading
to invalid results being returned from the profiler.

This fix ensures that resetting will reinitialize the profiler if it
is still enabled.