I think I fixed MacOS zlib!
I was finally able to get the MacOS builds to crash on my local machine, leading me to a likely fix for the issue. This issue was severe and caused the MacOS NVGT builds both compiler and compiled to be completely defunct on MacOS as zlib, the library used for compressing bytecode, certain hashing and other things, was not working on this platform. Future readers, sorry for the below spam.