front() and back() helpers for array (#273)
These are helpers (front() and back()) that retrieve the first and last elements respectively, they will throw the standard out of bounds exception if the array is empty.
These are functionally the same as array[0] and array[array.length()-1] or array[-1], but can make code more readable and intentions clearer if using an array as a queue or similar structure.