tts_voice major refactor (in progress)
This entirely reworks the tts_voice object internally. the nvgt API is still the same, but both code and feature wise the tts system is much cleaner, more extensible, more feature ritch and fewer bugs, more modern etc. Note to bleeding edge testers, this commit is being made for colaberative purposes and not because the system is 100% complete. Testing is welcome and appreciated so long as it's kept in mind that this is not production ready. The commit that completes the system will include a much more detailed feature list.